Lalonde Dataset

This dataset was used by Deheija and Wahba (1999) to evaluate propensity score matching. It was then used by Jasjeet Sekhon in the Matching package. It contains information pertaining to a range of demographic variables, real earnings, indicator for real earings being zero, and a treatment indicator, and is commonly used to demonstrate causal inference and analysis. The lalonde dataset consists of 445 observations and 12 variables, including:

age

age in years

educ

(number of years of education completed)

black

dummy variable to indicate if a person is black or not

hisp

dummy variable to indicate if a person is hispanic or not

married

dummy variable to indicate if a person is married or not

nodegr

dummy variable to indicate if a person has attained a highschool diploma or not

re74

real earnings in 1974

re75

real earnings in 1975

re78

real earnings in 1978

u74

dummy variable to indicate whether of not real earnings in 1974 were 0 or not

u75

dummy variable to indicate whether of not real earnings in 1975 were 0 or not

treat

treatment indicator

data(lalonde)

Format

An object of class rda

Source

Matching package

References

Dehejia, Rajeev and Sadek Wahba. 1999.“Causal Effects in Non-Experimental Studies: Re-Evaluating the Evaluation of Training Programs.” Journal of the American Statistical Association 94 (448): 1053-1062.