Summary.qed_match: Summarazing Matching statistics

Summary method for class "qed_match", which provides relevant pre-matched and post-matched summary statistics.

# S3 method for qed_match
summary(qed_match, ...)

Arguments

qed_match
an object of class "qed_match", a result of a call to qed_match (which is a wrapper for Match)

Value

The function summary.qed_match computes and prints summary statistics for data that Greedy Matching has been performed on using the qed_match function. Summary statistics are computed using the CreateTableOne function in the tableone package, and summary.Match (which is a summary method for class "Match") function in the Matching package.

Details

Standardized Mean Differences are the differences in means between groups (i.e. treatment and control), divided by the (pooled) standard deviation. Rules of thumb:

SMD < 0.1

indicates adequate balance

SMD 0.1-0.2

indicates moderate balance

SMD > 0.2

indicates imbalance

See also

Examples

if (FALSE) {
data(lalonde)
# For matching with propensity scores:
ps_match <- qed_match(treat ~ age + educ + black, "re78", lalonde, "propensityscore", caliper = NULL, replace = FALSE)
# For matching with Mahalanobis distances:
md_match <- qed_match(treat ~ age + educ + black, "re78", lalonde, "mahalanobis", caliper = NULL, replace = FALSE)

# Summary statistics:
summary(ps_match)
summary(md_match)
}