Generic Method for Outcome Analysis for Matched data (to estimate treatment effects)

Testing method for object of class "qed_match".

test(x, ...)

Arguments

qed_match
an object of class "qed_match", a result of a call to qed_match (which is a wrapper for Match)

Value

The function test.qed_match computes and prints outcome analysis to estimate treatment effects for data that Greedy Matching has been performed on using the qed_match function. It also provides an easy interpretation of test results.

Details

For numeric outcomes, we perform a t-test to check for significant differences in means between groups. For categorical outcomes, we perform a McNemar test to check for odds ratios for the treatment outcomes and control outcomes.

See also

Examples

data(lalonde)
# For matching with propensity scores and a categorical output:
lalonde$nodegr <- factor(lalonde$nodegr)
ps_match <- qed_match(treat ~ age + educ + black, "nodegr", lalonde, "propensityscore", caliper = NULL, replace = FALSE)
# For matching with Mahalanobis distances and a numeric output:
md_match <- qed_match(treat ~ age + educ + black, "re78", lalonde, "mahalanobis", caliper = NULL, replace = FALSE)

# Outcome Analysis
# McNemar test for categorical outcomes:
test(ps_match)

#> ---------------------------------------
#>     Categorical Outcome: nodegr
#> ---------------------------------------
#> 
#> 
#> 	Exact McNemar test (with central confidence intervals)
#> 
#> data:  table(con, trt)
#> b = 12, c = 26, p-value = 0.03355
#> alternative hypothesis: true odds ratio is not equal to 1
#> 95 percent confidence interval:
#>  0.2121584 0.9475223
#> sample estimates:
#> odds ratio 
#>  0.4615385 
#> 
#> --------------------------------------- 
#> The association is statistically significant at the 0.05 level. (p-value = 0.03355244)
#> Observations in the treated group are 0.4615385 as likely to be a positive case* 
#> *(i.e. belong to outcome (nodegr) level = 1).

# T-test for numeric outcomes:
test(md_match)

#> ---------------------------------------
#>     Numeric Outcome: re78
#> ---------------------------------------
#> 
#> 	Paired t-test
#> 
#> data:  con and trt
#> t = -2.9135, df = 184, p-value = 0.004017
#> alternative hypothesis: true difference in means is not equal to 0
#> 95 percent confidence interval:
#>  -3359.937  -646.740
#> sample estimates:
#> mean of the differences 
#>               -2003.338 
#> 
#> --------------------------------------- 
#> The association is statistically significant at the 0.01 level. (p-value = 0.004016937)
#> On average, observations in the treated group are -2003.338 units below the control group for the outcome (re78).